// src/lib/api/generated/model/v1NamespaceShareRelation.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere Hub API
 * OpenAPI spec version: v1
 */

export type V1NamespaceShareRelation = typeof V1NamespaceShareRelation[keyof typeof V1NamespaceShareRelation];


export const V1NamespaceShareRelation = {
  NAMESPACE_SHARE_RELATION_UNSPECIFIED: 'NAMESPACE_SHARE_RELATION_UNSPECIFIED',
  NAMESPACE_SHARE_RELATION_VIEWER: 'NAMESPACE_SHARE_RELATION_VIEWER',
  NAMESPACE_SHARE_RELATION_USER: 'NAMESPACE_SHARE_RELATION_USER',
  NAMESPACE_SHARE_RELATION_EDITOR: 'NAMESPACE_SHARE_RELATION_EDITOR',
} as const;